requirements.txt
pandas==2.*
minio==7.2.4
requests==2.31.0
load-dotenv==0.1.0
typer[all]==0.9.0
pytest==7.4.3
python-dotenv~=1.0.1
rich~=13.7.0
pyarrow~=15.0.0